fix for bug where syncVersion was shared across all rootDirs, refactor to make rootDir int instead of long to match db type

+++ common-sync/src/main/java/com/fstojilj/luddite/sync/common/model/FileMetadata.java
@@ -8,7 +8,7 @@ import java.time.Instant;
 public record FileMetadata(
         Long id,
         String filename,
-        Long rootDirId,
+        Integer rootDirId,
         String relativePath,
         Long fileSize,
         String checksum,
